test: librados startup/shutdown racer test
Spawn a bunch of threads, have them do a cluster_connect and immediately
shut the connection down again. Do this in a loop and allow the thread
to exit. Join all of the threads.

This helps ferret out places where state that is global to the program
is inadequately protected.

Test-for: https://tracker.ceph.com/issues/42026
Signed-off-by: Jeff Layton <jlayton@redhat.com>